Replace int by uint64_t on setNextTick definition, for compiling on 32 bit platforms (ARM and others).

+++ src/midilfo.cpp
@@ -605,7 +605,7 @@ void MidiLfo::applyPendingParChanges()
     needsGUIUpdate = true;
 }
 
-void MidiLfo::setNextTick(int tick)
+void MidiLfo::setNextTick(uint64_t tick)
 {
     int tickres = TPQN/res;
     int pos = (tick/tickres) % nPoints;
